Centralize and de-duplicate comparison and arith tests (#19317)
