+++ pandas/tests/scalar/test_timedelta.py
@@ -276,6 +276,14 @@ class TestTimedeltaComparison(object):
         assert res.shape == expected.shape
         assert (res == expected).all()
 
+    def test_compare_timedelta_ndarray(self):
+        # GH11835
+        periods = [Timedelta('0 days 01:00:00'), Timedelta('0 days 01:00:00')]
+        arr = np.array(periods)
+        result = arr[0] > arr
+        expected = np.array([False, False])
+        tm.assert_numpy_array_equal(result, expected)
+
 
 class TestTimedeltas(object):
     _multiprocess_can_split_ = True
